fix: lazy-import qt backend and extract ansi_to_html to avoid Qt ModuleNotFoundError in CI
The qt/__init__.py eagerly imported Qt via `from .main import UserInterface`,
causing test collection to fail in headless CI environments without Qt.

- Replace eager import in qt/__init__.py with PEP 562 lazy __getattr__
  (matches the pattern already applied to tk/__init__.py in PR #15)
- Extract ansi_to_html(), ANSI_COLOR_MAP, and ansi_color_re from
  qt/process_output_widget.py into new ros2launch_gui/ansi.py module
  (pure Python, zero GUI dependencies)
- Update test_ansi_to_html.py to import from ros2launch_gui.ansi

Closes #11
