+++ ros2launch_gui/qt/__init__.py
@@ -1,5 +1,8 @@
-from .main import UserInterface
+__all__ = ['UserInterface']
 
-__all__ = [
-    'UserInterface',
-]
+
+def __getattr__(name):
+    if name == 'UserInterface':
+        from .main import UserInterface
+        return UserInterface
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
